requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
astronomer-cosmos
apache-airflow-providers-snowflake
asn1crypto==1.5.1
bleach==6.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cramjam==2.9.1
cryptography==44.0.0
fastparquet==2024.11.0
filelock==3.17.0
fsspec==2025.2.0
apache-airflow
idna==3.10
kaggle==1.6.17
numpy==2.2.2
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
pyarrow==19.0.0
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2025.1
requests==2.32.3
six==1.17.0
snowflake-connector-python==3.13.2
sortedcontainers==2.4.0
text-unidecode==1.3
tomlkit==0.13.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webencodings==0.5.1
astro-run-dag 
apache-airflow-providers-snowflake